[#49] 🚚 rename modal contents directory to modal content + add a module bridge
