+++ src/components/shared/modalContent/index.ts
@@ -1,0 +1,3 @@
+import { ModalContent } from './ModalContent'
+
+export { ModalContent }
